% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are.colors}
\alias{are.colors}
\title{are_colors   
Helper function to check for valid color strings}
\usage{
are.colors(x)
}
\arguments{
\item{x}{color vector (etc) to test.}
}
\description{
are_colors   
Helper function to check for valid color strings
}
