% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{d3heatmap-exports}
\alias{d3heatmap-exports}
\alias{\%>\%}
\title{d3heatmap exported operators and S3 methods 
 
The following functions are imported and then re-exported 
from the d3heatmap package to avoid listing the various packages
as Depends of d3heatmap}
\description{
d3heatmap exported operators and S3 methods 
 
The following functions are imported and then re-exported 
from the d3heatmap package to avoid listing the various packages
as Depends of d3heatmap
}
