% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3heatmap.R
\name{d3heatmap}
\alias{d3heatmap}
\title{D3 Heatmap widget}
\usage{
d3heatmap(data, theme = "", colors = "RdYlBu", width = NULL,
  height = NULL, cluster = !any(is.na(data)))
}
\description{
Creates a D3.js-based heatmap widget.
}

