% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\docType{class}
\name{d3heatmapGadget-class}
\alias{d3heatmapGadget-class}
\title{An S4 class to capture a heatmap}
\description{
An S4 class to capture a heatmap
}
\section{Slots}{

\describe{
\item{\code{x}}{A data.frame}

\item{\code{ColSideColors}}{(optional) character vector of length ncol(x), or matrix with
columns equal to ncol(x), containing the color names for a horizontal side bar that may 
be used to annotate the columns of x. \code{ColIndividualColors}, from heatmap.2, can
also be used}

\item{\code{RowSideColors}}{(optional) character vector of length nrow(x), or matrix with rows
equal to nrow(x), containing the color names for a vertical side bar that may be used to
annotate the rows of x. \code{RowIndividualColors}, from heatmap.3, can also be used}

\item{\code{settings}}{A list of settings to pass to d3heatmap()}

\item{\code{heatmap}}{A d3heatmap}

\item{\code{rows}}{Character vector of selected rows from the heatmap}

\item{\code{cols}}{Character vector of selected columns from the heatmap}

\item{\code{filter}}{Character string of a valid filter (as in \code{subset()} for the underlying 
d3heatmap data; invalid filters will be safely loaded but not applied}
}}

