% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{d3heatmapGadget}
\alias{d3heatmapGadget}
\title{A shiny gadget for d3heatmaps}
\usage{
d3heatmapGadget(x, ...)
}
\arguments{
\item{x}{A numeric matrix or data.frame populated with all numerics, or a list 
of class \emph{d3heatmapGadget}}

\item{...}{arguments to be passed to \code{d3heatmap()}}
}
\value{
An object of class \emph{d3heatmapGadget} containing the original data,
gadget settings, filter parameters, and the resulting plot
}
\description{
Creates a D3.js-based heatmap widget inside a Shiny gadget. Pass into the function 
either a d3heatmap-compatible matrix or data.frame, or a list of class 
\emph{d3heatmapGadget} to start from the gadget's last saved settings
}
\examples{
\dontrun{

hmg <- d3heatmapGadget(mtcars)

class(hmg)
# "d3heatmapGadget"

}

}
