% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3heatmap.R
\name{d3heatmapOutput}
\alias{d3heatmapOutput}
\title{Widget output function for use in Shiny}
\usage{
d3heatmapOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

