% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmColors.R
\name{hmColors}
\alias{hmColors}
\title{Set d3heatmap color options}
\source{
The interface was inspired by \link{dygraphs}
}
\usage{
hmColors(d3heatmap, theme = c("", "dark"), colors, range, color.bins,
  symmetrical, na.color)
}
\arguments{
\item{d3heatmap}{\emph{Required} A valid \emph{d3heatmap} object}

\item{theme}{A custom CSS theme to use. Currently the only valid values 
are \code{""} and \code{"dark"}. \code{"dark"} is primarily intended for 
  standalone visualizations, not R Markdown or Shiny.}

\item{colors}{Either a colorbrewer2.org palette name (e.g. 
\code{"YlOrRd"} or \code{"Blues"}), or a vector of colors 
to interpolate in hexadecimal \code{"#RRGGBB"} format, or a color 
interpolation function like \code{\link[grDevices]{colorRamp}}.}

\item{na.color}{Color of NA values in heatmap. Defaults to neutral gray.}

\item{bins}{\emph{integer} The number of colors to generate from 
the palette.}

\item{symbreaks}{\emph{logical} Arrange color bins symmetrically around 
zero?}

\item{rng}{A vector of two numbers, namely the minimum and maximum value
  to use when determining the mapping from values to colors. This is 
useful when the range of values changes between heatmaps, but colors 
should be the same (optional, defaults to the minimum and maximum 
of \code{x}).}
}
\description{
Set and adjust the colors and color options for the d3heatmap
}
