% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmDendrogram.R
\name{hmDendrogram}
\alias{hmDendrogram}
\title{Set d3heatmap dendrograms}
\source{
The interface was inspired by \cite{dygraphs}
}
\usage{
hmDendrogram(
  d3heatmap,
  dendrogram = c("row", "column", "both", "none"),
  reorder,
  row.reorder,
  column.reorder,
  distance.function,
  clustering.function,
  reorder.function,
  groups,
  symmetrical
)
}
\arguments{
\item{d3heatmap}{\emph{Required} A valid \emph{d3heatmap} object}

\item{dendrogram}{\emph{Required} The dendrogram to process in this
call, among \code{'row', 'col', 'both', 'none'}. Using the \code{'none'}
value turns off any previous dendrogram settings. If reorder or row.reorder 
and col.reorder are FALSE or NULL and dendrogram is 'both', then a warning 
is issued and row.reorder (or col.reorder) arguments are honored.}

\item{reorder}{a parameter that allows the user to pass in a single 
reordering value to use on the dendrograms specified in the 
\code{dendrogram} argument.}

\item{row.reorder}{determines if and how the row dendrogram 
should be reordered.    By default, it is TRUE, which implies 
dendrogram is computed and reordered based on row means. 
If NULL or FALSE, then no dendrogram is computed and no reordering 
is done. If a dendrogram, then it is used "as-is", i.e., without any 
reordering. If a vector of integers, then dendrogram is computed 
and reordered based on the order of the vector.}

\item{column.reorder}{determines if and how the column dendrogram should be 
be reordered.    Has the options as the Rowv argument above and additionally 
when x is a square matrix, Colv = "Rowv" means that columns should be 
treated identically to the rows.}

\item{distance.function}{function used to compute the distance 
(dissimilarity) between both rows and columns. Defaults to dist.}

\item{clustering.function}{function used to compute the hierarchical 
clustering when Rowv or Colv are not dendrograms. Defaults to hclust.}

\item{reorder.function}{function(d, w) of dendrogram and weights for 
reordering the row and column dendrograms.}

\item{groups}{an integer scalar with the desired number of groups by which 
to color the dendrogram's branches (uses \link[dendextend]{color_branches})}

\item{symmetrical}{logical indicating if x should be treated symmetrically; can only be true when x is a square matrix.}
}
\value{
Modified d3heatmap object
}
\description{
Set and adjust the dendrograms for a d3heatmap object
}
\examples{
\dontrun{

d3heatmap(mtcars, scale = "column", col = "Blues") \%>\%
  hmDendrogram(dendrogram = 'row', groups = 3)

}

}
\seealso{
\link{heatmap}, \link[gplots]{heatmap.2}
}
