% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmOptions.R
\name{hmOptions}
\alias{hmOptions}
\title{Modify d3heatmap options}
\source{
The interface was inspired by \cite{dygraphs}
}
\usage{
hmOptions(d3heatmap, title, scale = c("none", "row", "column"),
  scale.by.range, na.rm, na.value, show.grid, animation.duration,
  reverse.columns)
}
\arguments{
\item{d3heatmap}{a d3heatmap object created from the d3heatmap()}

\item{title}{\emph{string} Plot title. Defaults to \code{NULL}.}

\item{scale}{character indicating if the values should be centered and scaled in either 
the row direction or the column direction, or none. The default is "none".}

\item{scale.by.range}{logical indicating whether to scale rows or columns by
the range of each row and column. Setting this parameter to \code{TRUE} 
automatically sets \code{na.rm} to \code{TRUE}}

\item{na.rm}{logical indicating whether NA's should be removed.}

\item{na.value}{numeric indicating where NA's should be substituted to trigger the NA color.}

\item{show.grid}{\code{TRUE} to show gridlines, \code{FALSE} to hide them, 
or a numeric value to specify the gridline thickness in pixels (can 
be a non-integer).}

\item{animation.duration}{Number of milliseconds to animate zooming in and 
out. For large \code{x} it may help performance to set this value to 
\code{0}.}

\item{reverse.columns}{logical indicating if the column order should be 
reversed for plotting. Default (when missing) - is FALSE, unless symmetrical 
is TRUE... useful for correlation matrices.}
}
\value{
Modified d3heatmap object
}
\description{
Change options for a d3heatmap
}
\examples{
library(d3heatmap)

d3heatmap(mtcars, scale = "column", col = "Blues") \%>\%
  hmOptions(scale.by.range = TRUE, show.grid = FALSE,
 		animation.duration = 400)

}
\seealso{
\link{heatmap}, \link[gplots]{heatmap.2}
}
