% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmSidecolors.R
\name{hmSideColors}
\alias{hmSideColors}
\title{Modify axis options}
\source{
The interface was inspired by \cite{dygraphs}
}
\usage{
hmSideColors(d3heatmap, axis = c("x", "y", "row", "column"), side.colors,
  palette, names)
}
\arguments{
\item{d3heatmap}{a d3heatmap object created from the d3heatmap()}

\item{axis}{Name of the axis to modify; either "x", "y", "row", or "column"}

\item{side.colors}{character vector of length ncol(x), or matrix with
columns equal to ncol(x), containing the color names for a horizontal 
side bar that may 
be used to annotate the rows or columns of x.}

\item{palette}{a palette of colors to use for side colors if passing a 
non-color matrix}

\item{names}{the names of the variables for the side colors.}
}
\description{
Provides options for modifying the x and y axes of a D3 Heatmap widget
}
\examples{
library(d3heatmap)

rsc <- matrix(rep_len(c('good', 'bad', 'ugly'), length.out = 64), ncol = 2)
rsccols <- c('red', 'white', 'blue')
rscnames <- c('Row 1', 'Row 2')

csc <- matrix(rep_len(c('first', 'second', 'third', 'fourth', 'fifth'), length.out = 33), nrow = 3)
csccols <- c('orange', 'blue', 'grey', 'green', 'red')
cscnames <- c('Column 1', 'Column 2', 'Column 3')

d3heatmap(mtcars, key = TRUE, scale = 'column') \%>\% 
  hmSideColors(axis = 'column', side.colors = csc,
    palette = csccols, names = cscnames) \%>\% 
  hmSideColors(axis = 'y', side.colors = rsc,
    palette = rsccols, names = rscnames)


}
\seealso{
\link{heatmap}, \link[gplots]{heatmap.2}
}
