% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,d3heatmapGadget-method}
\alias{print.d3heatmapGadget}
\title{Print a gadget's main output}
\usage{
print(gadget)

\S4method{print}{d3heatmapGadget}(gadget)
}
\arguments{
\item{gadget}{The gadget to print to screen}
}
\description{
Use this method to print the main component of the d3heatmapGadget to the viewer 
pane without intermediate code
}
\examples{
\dontrun{

hm <- d3heatmapGadget(mtcars)
print(hm)

}


}
