% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\docType{methods}
\name{save}
\alias{save}
\alias{save,d3heatmapGadget-method}
\alias{save.d3heatmapGadget}
\title{Save a gadget's main output}
\usage{
save(gadget, ...)

\S4method{save}{d3heatmapGadget}(gadget, file = NULL, ...)
}
\arguments{
\item{gadget}{The gadget to save to file}

\item{...}{arguments to pass to \code{htmlwidgets::saveWidget}}

\item{file}{the name of the file, without extension}
}
\description{
Use this method to save the d3heatmap from the d3heatmapGadget
without intermediary code.
}
\examples{
\dontrun{

hm <- d3heatmapGadget(mtcars)
save(hm, file = "myHeatmap")

}

}
