% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3heatmap.R
\name{d3heatmap}
\alias{d3heatmap}
\title{D3 Heatmap widget}
\source{
The interface was designed based on \link{heatmap} and \link[gplots]{heatmap.2}
}
\usage{
d3heatmap(x, main = NULL, width = NULL, height = NULL, show_grid = TRUE,
  anim_duration = 500, rng = NULL, Rowv = TRUE, Colv = if (symm) "Rowv"
  else TRUE, distfun = dist, hclustfun = hclust, dendrogram = c("both",
  "row", "column", "none"), reorderfun = function(d, w) reorder(d, w),
  k_row = NULL, k_col = NULL, symm = FALSE, revC = NULL,
  scale = c("none", "row", "column"), na.rm = TRUE, na.color = "#777777",
  na.value = NA, legend.title = NULL, show.legend = FALSE,
  legend.location = c("fl", "br", "tr", "tl", "tr"), digits = 3L,
  cellnote = NULL, cellnote_scale = FALSE, cellnote_row = NULL,
  cellnote_col = NULL, cellnote_val = "Value", brush_color = "#0000FF",
  print.values = FALSE, theme = NULL, colors = "RdYlBu", bins = NULL,
  symbreaks = FALSE, labRow = rownames(x), labCol = colnames(x),
  xaxis_height = 80, yaxis_width = 120, xaxis_font_size = NULL,
  yaxis_font_size = NULL, xaxis_angle = 60, xaxis_location = "bottom",
  yaxis_location = "right", xaxis_title = NULL, yaxis_title = NULL,
  xaxis_title_font_size = 14, yaxis_title_font_size = 14, cexRow = NULL,
  cexCol = NULL)
}
\arguments{
\item{x}{A numeric matrix
Defaults to \code{TRUE} unless \code{x} contains any \code{NA}s.}

\item{main}{\emph{string} Plot title. Defaults to \code{NULL}.}

\item{width}{Width in pixels (optional, defaults to automatic sizing).}

\item{height}{Height in pixels (optional, defaults to automatic sizing).}

\item{show_grid}{\code{TRUE} to show gridlines, \code{FALSE} to hide them, or
a numeric value to specify the gridline thickness in pixels (can be a non-integer).}

\item{anim_duration}{Number of milliseconds to animate zooming in and out.
For large \code{x} it may help performance to set this value to \code{0}.}

\item{rng}{A vector of two numbers, namely the minimum and maximum value
to use when determining the mapping from values to colors. This is useful
when the range of values changes between heatmaps, but colors should be the
same (optional, defaults to the minimum and maximum of \code{x}).}

\item{Rowv}{determines if and how the row dendrogram 
should be reordered.    By default, it is TRUE, which implies 
dendrogram is computed and reordered based on row means. 
If NULL or FALSE, then no dendrogram is computed and no reordering 
is done. If a dendrogram, then it is used "as-is", ie without any 
reordering. If a vector of integers, then dendrogram is computed 
and reordered based on the order of the vector.}

\item{Colv}{determines if and how the column dendrogram should be 
reordered.    Has the options as the Rowv argument above and additionally 
when x is a square matrix, Colv = "Rowv" means that columns should be 
treated identically to the rows.}

\item{distfun}{function used to compute the distance (dissimilarity) 
between both rows and columns. Defaults to dist.}

\item{hclustfun}{function used to compute the hierarchical clustering 
when Rowv or Colv are not dendrograms. Defaults to hclust.}

\item{dendrogram}{character string indicating whether to draw 'none', 
'row', 'column' or 'both' dendrograms. Defaults to 'both'. However, if 
Rowv (or Colv) is FALSE or NULL and dendrogram is 'both', then a warning 
is issued and Rowv (or Colv) arguments are honoured.}

\item{reorderfun}{function(d, w) of dendrogram and weights for reordering 
the row and column dendrograms. The default uses stats{reorder.dendrogram}}

\item{k_row}{an integer scalar with the desired number of groups by which 
to color the dendrogram's branches in the rows 
(uses \link[dendextend]{color_branches})}

\item{k_col}{an integer scalar with the desired number of groups by which 
to color the dendrogram's branches in the columns 
(uses \link[dendextend]{color_branches})}

\item{symm}{logical indicating if x should be treated symmetrically; can only be true when x is a square matrix.}

\item{revC}{logical indicating if the column order should be reversed for plotting.
Default (when missing) - is FALSE, unless symm is TRUE.
This is useful for cor matrix.}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "none".}

\item{na.rm}{logical indicating whether NA's should be removed.}

\item{na.color}{Color of NA values in heatmap. Defaults to neutral gray.}

\item{na.value}{numeric indicating where NA's should be substituted to trigger the NA color.}

\item{legend.title}{Separate title for legend. Defaults to \code{NULL}.}

\item{show.legend}{Show color key and density
information? \code{TRUE/FALSE}. Defaults to \code{FALSE}}

\item{legend.location}{\emph{Required, character} Location for legend, either \code{"fl", "br", "tr"
"tl", or "bl"} for "float", "bottom right", "top right", "top left", and "bottom left". 
Defaults to \code{"fl"}, which follows the location of the axis labels.}

\item{digits}{\emph{integer} indicating the number of decimal places to be used by \link{round} for 'label'.}

\item{cellnote}{\emph{numeric} (optional) matrix of the same dimensions as \code{x} that has the human-readable version of each value, for displaying to the user on hover. If \code{NULL}, then \code{x} will be coerced using \code{\link{as.character}}.
If missing, it will use \code{x}, after rounding it based on the \code{digits} parameter.}

\item{cellnote_scale}{\emph{logical} (default is FALSE). IF cellnote is missing and x is used, 
should cellnote be scaled if x is also scaled?}

\item{cellnote_row}{\emph{character} Label to display next to the row value when the user hovers over the cell.
If not specified, trys to match the \code{xaxis_title}; if no axis title, defaults to "Row".}

\item{cellnote_col}{\emph{character} Label to display next to the column value when the user hovers over the cell. 
If not specified, trys to match the \code{yaxis_title}; if no axis title, defaults to "Col".}

\item{cellnote_val}{\emph{character} Label to display next to the cell value when the user hovers over the cell.
Defaults to "Value".}

\item{brush_color}{The base color to be used for the brush. The brush will be
filled with a low-opacity version of this color. \code{"#RRGGBB"} format 
expected.}

\item{print.values}{\emph{logical} Show the values inside the cells. Defatuls to \code{FALSE}.}

\item{theme}{A custom CSS theme to use. Currently the only valid values are 
\code{""} and \code{"dark"}. \code{"dark"} is primarily intended for 
standalone visualizations, not R Markdown or Shiny.}

\item{colors}{Either a colorbrewer2.org palette name (e.g. \code{"YlOrRd"} 
or \code{"Blues"}), or a vector of colors to interpolate in hexadecimal 
\code{"#RRGGBB"} format, or a color interpolation function like
\code{\link[grDevices]{colorRamp}}.}

\item{bins}{\emph{integer} The number of colors to generate from the palette.}

\item{symbreaks}{\emph{logical} Arrange color bins symmetrically around zero?}

\item{labRow}{character vectors with row labels to use (from top to bottom); default to rownames(x).}

\item{labCol}{character vectors with column labels to use (from left to right); default to colnames(x).}

\item{xaxis_height}{Size of axes, in pixels.}

\item{yaxis_width}{Size of axes, in pixels.}

\item{xaxis_font_size}{Font size of axis labels, as a CSS size (e.g. "14px" or "12pt").}

\item{yaxis_font_size}{Font size of axis labels, as a CSS size (e.g. "14px" or "12pt").}

\item{xaxis_angle}{Angle of axis labels. Defaults to 60. Maximum of 90 (vertical), minimum of 25.}

\item{xaxis_location}{Location of the x axis, either "bottom" or "top". Defaults to "bottom".}

\item{yaxis_location}{Location of the y axis, either "right" or "left". Defaults to "right".}

\item{xaxis_title}{Title text of x axis}

\item{yaxis_title}{Title text of y axis}

\item{xaxis_title_font_size}{Font size of x axis title in pixels. Defaults to 14.}

\item{yaxis_title_font_size}{Font size of y axis title in pixels. Defaults to 14.}

\item{cexRow}{positive numbers. If not missing, it will override \code{xaxis_font_size}
and will give it a value cexRow*14}

\item{cexCol}{positive numbers. If not missing, it will override \code{yaxis_font_size}
and will give it a value cexCol*14}
}
\description{
Creates a D3.js-based heatmap widget.
}
\examples{
library(d3heatmap)
d3heatmap(mtcars, scale = "column", colors = "Blues")


}
\seealso{
\link{heatmap}, \link[gplots]{heatmap.2}
}
