% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3heatmap.R
\name{d3heatmap}
\alias{d3heatmap}
\title{D3 Heatmap widget}
\source{
The interface was designed based on \link{heatmap} and \link[gplots]{heatmap.2}
}
\usage{
d3heatmap(x, main = NULL, width = NULL, height = NULL, show_grid = TRUE,
  anim_duration = 500, rng = NULL, symm = FALSE, Rowv = TRUE,
  Colv = if (symm) "Rowv" else TRUE, distfun = dist, hclustfun = hclust,
  dendrogram = c("both", "row", "column", "none"), reorderfun = function(d,
  w) reorder(d, w), k_row = NULL, k_col = NULL, kr, kc, revC = NULL,
  scale = c("none", "row", "column"), scale.by.range = FALSE,
  na.rm = TRUE, na.color = "#777777", na.value = NA, key = FALSE,
  keysize = 1, key.title = NULL, key.location = c("fl", "br", "tr", "tl",
  "bl"), density.info = c("histogram", "none"), denscol = NULL,
  digits = 3L, cellnote = NULL, cellnote_scale = FALSE,
  cellnote_row = NULL, cellnote_col = NULL, cellnote_val = "Value",
  brush_color = "#0000FF", print.values = FALSE, notecol = "#222222",
  notecex = 1, cex.note, theme = NULL, col = "RdYlBu", breaks = NULL,
  symbreaks = FALSE, labRow = rownames(x), labCol = colnames(x),
  labColSize = 80, labRowSize = 120, cexCol = NULL, cexRow = NULL,
  srtCol = 60, sideCol = 3, sideRow = 4, xlab = NULL, ylab = NULL,
  xaxis_title_font_size = 14, yaxis_title_font_size = 14,
  ColSideColors = NULL, RowSideColors = NULL, RowColorsPalette = NULL,
  ColColorsPalette = NULL, RowColorsNames = NULL, ColColorsNames = NULL,
  ...)
}
\arguments{
\item{x}{A numeric matrix
Defaults to \code{TRUE} unless \code{x} contains any \code{NA}s.}

\item{main}{\emph{string} Plot title. Defaults to \code{NULL}.}

\item{width}{Width in pixels (optional, defaults to automatic sizing).}

\item{height}{Height in pixels (optional, defaults to automatic sizing).}

\item{show_grid}{\code{TRUE} to show gridlines, \code{FALSE} to hide them, or
a numeric value to specify the gridline thickness in pixels (can be a non-integer).}

\item{anim_duration}{Number of milliseconds to animate zooming in and out.
For large \code{x} it may help performance to set this value to \code{0}.}

\item{rng}{A vector of two numbers, namely the minimum and maximum value
to use when determining the mapping from values to colors. This is useful
when the range of values changes between heatmaps, but colors should be the
same (optional, defaults to the minimum and maximum of \code{x}).}

\item{symm}{logical indicating if x should be treated symmetrically; can only be true when x is a square matrix.}

\item{Rowv}{determines if and how the row dendrogram 
should be reordered.    By default, it is TRUE, which implies 
dendrogram is computed and reordered based on row means. 
If NULL or FALSE, then no dendrogram is computed and no reordering 
is done. If a dendrogram, then it is used "as-is", ie without any 
reordering. If a vector of integers, then dendrogram is computed 
and reordered based on the order of the vector.}

\item{Colv}{determines if and how the column dendrogram should be 
reordered.    Has the options as the Rowv argument above and additionally 
when x is a square matrix, Colv = "Rowv" means that columns should be 
treated identically to the rows.}

\item{distfun}{function used to compute the distance (dissimilarity) 
between both rows and columns. Defaults to dist.}

\item{hclustfun}{function used to compute the hierarchical clustering 
when Rowv or Colv are not dendrograms. Defaults to hclust.}

\item{dendrogram}{character string indicating whether to draw 'none', 
'row', 'column' or 'both' dendrograms. Defaults to 'both'. However, if 
Rowv (or Colv) is FALSE or NULL and dendrogram is 'both', then a warning 
is issued and Rowv (or Colv) arguments are honoured.}

\item{reorderfun}{function(d, w) of dendrogram and weights for reordering 
the row and column dendrograms. The default uses stats{reorder.dendrogram}}

\item{k_row, kr}{an integer scalar with the desired number of groups by which 
to color the dendrogram's branches in the rows 
(uses \link[dendextend]{color_branches})}

\item{k_col, kc}{an integer scalar with the desired number of groups by which 
to color the dendrogram's branches in the columns 
(uses \link[dendextend]{color_branches})}

\item{revC}{logical indicating if the column order should be reversed for plotting.
Default (when missing) - is FALSE, unless symm is TRUE.
This is useful for cor matrix.}

\item{scale}{character indicating if the values should be centered and scaled in either 
the row direction or the column direction, or none. The default is "none".}

\item{scale.by.range}{logical indicating whether to scale rows or columns by
the range of each row and column. Setting this parameter to \code{TRUE} 
automatically sets \code{na.rm} to \code{TRUE}}

\item{na.rm}{logical indicating whether NA's should be removed.}

\item{na.color}{Color of NA values in heatmap. Defaults to neutral gray.}

\item{na.value}{numeric indicating where NA's should be substituted to trigger the NA color.}

\item{key}{Show color key and density
information? \code{TRUE/FALSE}. Defaults to \code{FALSE}}

\item{keysize}{\emph{numeric} value indicating the relative size of the key. Will multiple the labRowSize and labColSize.
Defaults to 1}

\item{key.title}{Separate title for color key. Defaults to \code{NULL}.}

\item{key.location}{\emph{Required, character} Location for color key, either \code{"fl", "br", "tr"
"tl", or "bl"} for "float", "bottom right", "top right", "top left", and "bottom left". 
Defaults to \code{"fl"}, which follows the location of the axis labels.}

\item{density.info}{\emph{Optional, character} indicating whether to superimpose 
a 'histogram' or no plot ('none') on the color-key (legend).}

\item{denscol}{\emph{Optional, character} giving the color for the density 
display specified by 'density.info', defaults to \code{NULL} which will fill the
bars with the same color as the bin}

\item{digits}{\emph{integer} indicating the number of decimal places to be used by \link{round} for 'label'.}

\item{cellnote}{\emph{numeric} (optional) matrix of the same dimensions as \code{x} that has the human-readable version of each value, for displaying to the user on hover. If \code{NULL}, then \code{x} will be coerced using \code{\link{as.character}}.
If missing, it will use \code{x}, after rounding it based on the \code{digits} parameter.}

\item{cellnote_scale}{\emph{logical} (default is FALSE). IF cellnote is missing and x is used, should cellnote be scaled if x is also scaled?}

\item{cellnote_row}{\emph{character} Label to display next to the row value when the user hovers over the cell.
If not specified, trys to match the \code{xlab}; if no axis title, defaults to "Row".}

\item{cellnote_col}{\emph{character} Label to display next to the column value when the user hovers over the cell. 
If not specified, trys to match the \code{ylab}; if no axis title, defaults to "Col".}

\item{cellnote_val}{\emph{character} Label to display next to the cell value when the user hovers over the cell.
Defaults to "Value".}

\item{brush_color}{The base color to be used for the brush. The brush will be
filled with a low-opacity version of this color. \code{"#RRGGBB"} format 
expected.}

\item{print.values}{\emph{logical} Show the values inside the cells. Defatuls to \code{FALSE}.}

\item{notecol}{\emph{character} name or hex specifying the color of the values printed
inside the cells}

\item{notecex, cex.note}{\emph{numeric} scalar of 12 to determine pixel size of printed cell values.
inside the cells. If not specified, the minimum font size between the x and y axes.}

\item{theme}{A custom CSS theme to use. Currently the only valid values are 
\code{""} and \code{"dark"}. \code{"dark"} is primarily intended for 
standalone visualizations, not R Markdown or Shiny.}

\item{col}{Either a colorbrewer2.org palette name (e.g. \code{"YlOrRd"} 
or \code{"Blues"}), or a vector of colors to interpolate in hexadecimal 
\code{"#RRGGBB"} format, or a color interpolation function like
\code{\link[grDevices]{colorRamp}}.}

\item{breaks}{\emph{integer} Either a numeric vector indicating the splitting 
points for binning x into colors, or a integer number of break points 
to be used, in which case the break points will be spaced equally 
between min(x) and max(x).}

\item{symbreaks}{\emph{logical} Arrange color bins symmetrically around zero?}

\item{labRow}{character vectors with row labels to use (from top to bottom); default to rownames(x).}

\item{labCol}{character vectors with column labels to use (from left to right); default to colnames(x).}

\item{labColSize}{Size of axes, in pixels.}

\item{labRowSize}{Size of axes, in pixels.}

\item{cexCol}{Font size of column labels, as a scalar to '12px' font. Default to 
0.2 + 1/log10(ncol(x)).}

\item{cexRow}{Font size of row labels, as a scalar to '12px' font size. Defaults to 
0.2 + 1/log10(nrow(x)).}

\item{srtCol}{Angle of axis labels. Defaults to 60. Maximum of 90 (vertical), minimum of 25.}

\item{sideCol}{3 or 1, for "bottom" or "top", which side column labels 
display. Defaults to 3 ("bottom").}

\item{sideRow}{2 or 4, for "left" or "right", which side row labels 
display. Defaults to 4 ("right").}

\item{xlab}{Title text of x axis}

\item{ylab}{Title text of y axis}

\item{xaxis_title_font_size}{Font size of x axis title in pixels. Defaults to 14.}

\item{yaxis_title_font_size}{Font size of y axis title in pixels. Defaults to 14.}

\item{ColSideColors}{(optional) character vector of length ncol(x), or matrix with
columns equal to ncol(x), containing the color names for a horizontal side bar that may 
be used to annotate the columns of x. \code{ColIndividualColors}, from heatmap.2, can
also be used}

\item{RowSideColors}{(optional) character vector of length nrow(x), or matrix with rows
equal to nrow(x), containing the color names for a vertical side bar that may be used to
annotate the rows of x. \code{RowIndividualColors}, from heatmap.3, can also be used}

\item{RowColorsPalette}{a palette of colors to use for RowSideColors if passing a non-color matrix}

\item{ColColorsPalette}{a palette of colors to use for ColSideColors if passing a non-color matrix}

\item{RowColorsNames}{the names of the variables for RowSideColors. Overrides colnames(RowSideColors)}

\item{ColColorsNames}{the names of the variables for ColSideColors. Overrides rownames(ColSideColors)
with rows equal to nrow(x), containing the color names for a vertical side bar that may be used to annotate the 
rows of x.}

\item{...}{a catch for undocumented features or un-used arguments from heatmap.2
or heatmap.3, to enable direct use of those formulations in d3heatmap}
}
\description{
Creates a D3.js-based heatmap widget.
}
\examples{
library(d3heatmap)
d3heatmap(mtcars, scale = "column", col = "Blues")

}
\seealso{
\link{heatmap}, \link[gplots]{heatmap.2}
}
