% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmAxis.R
\name{hmAxis}
\alias{hmAxis}
\title{Modify axis options}
\source{
The interface was inspired by \link{dygraphs}
}
\usage{
hmAxis(d3heatmap, axis = c("x", "y", "row", "column"), size = NULL, labels,
  font.size = NULL, angle = NULL, location, title = NULL,
  title.font.size = NULL)
}
\arguments{
\item{axis}{Name of the axis to modify; either "x", "y", "row", or "column"}

\item{size}{Size of axes, in pixels.}

\item{labels}{character vectors with axis labels to use (top to bottom for y axis, left to right for x); default to rownames(x) or colnames(x).}

\item{font.size}{Font size of axis labels, as a CSS size (e.g. "14px" or "12pt").}

\item{angle}{Angle of x axis labels (x axis only). Defaults to 60. Maximum of 90 (vertical), minimum of 25.}

\item{location}{Location of the axis, either "bottom" or "top" for the x axis, and 
either "right" or "left" for the y axis. Defaults to "bottom" and "right".}

\item{title}{Title text}

\item{title.font.size}{Font size of axis title in pixels. Defaults to 14.}

\item{heatmap}{a d3heatmap object created from the d3heatmap() or hmHeatmap()}
}
\value{
Modified d3heatmap object
}
\description{
Provides options for modifying the x and y axes of a D3 Heatmap widget
}
\examples{
library(d3heatmap)
library(dplyr)
d3heatmap(mtcars, scale = "column", colors = "Blues") \%>\%
  hmAxis("x", angle = 30, title = "test", location = 'top', font.size = '24px') \%>\% 
  hmAxis("y", title = "test", location = 'right')

}
\seealso{
\link{heatmap}, \link[gplots]{heatmap.2}
}
