% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmCells.R
\name{hmCells}
\alias{hmCells}
\title{Modify cell options}
\source{
The interface was inspired by \link{dygraphs}
}
\usage{
hmCells(d3heatmap, digits, cellnotes, font.size, color, scale, row.label,
  col.label, value.label, brush.color, print)
}
\arguments{
\item{d3heatmap}{\emph{Required} A valid \emph{d3heatmap} object}

\item{digits}{\emph{integer} (Optional) indicating the number of decimal places to be used by \link{round} for 'label'.}

\item{cellnotes}{\emph{numeric} (Optional) matrix of the same dimensions as \code{x} that has the human-readable version of each value, for displaying to the user on hover. If \code{NULL}, then \code{x} will be coerced using \code{\link{as.character}}.
If missing, it will use \code{x}, after rounding it based on the \code{digits} parameter.}

\item{font.size}{\emph{numeric} the pixel size of printed cell}

\item{color}{\emph{character} name or hex specifying the color of the values printed
inside the cells}

\item{scale}{\emph{logical} (default is FALSE). If cellnote is missing and x is used, should cellnotes be scaled if x is also scaled?}

\item{row.label}{\emph{character} (Optional) Label to display next to the row value when the user hovers over the cell.
If not specified, trys to match the \code{xaxis_title}; if no axis title, defaults to "Row".}

\item{col.label}{\emph{character} (Optional) Label to display next to the column value when the user hovers over the cell
If not specified, trys to match the \code{yaxis_title}; if no axis title, defaults to "Col".}

\item{value.label}{\emph{character} (Optional) Label to display next to the cell value when the user hovers over the cell.
Defaults to "Value".}

\item{print}{\emph{logical} Show the values inside the cells. Defatuls to \code{FALSE}.}
}
\value{
Modified d3heatmap object
}
\description{
Provides options for modifying the cells and the display of their values
}
\examples{
library(d3heatmap)
library(dplyr)
d3heatmap(mtcars, scale = "column", col = "Blues") \%>\%
  hmCells(digits = 0L, print = T)

}
\seealso{
\link{heatmap}, \link[gplots]{heatmap.2}
}
