% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmLegend.R
\name{hmLegend}
\alias{hmLegend}
\title{Modify the d3heatmap legend}
\usage{
hmLegend(heatmap, show = FALSE, title = NULL, location = c("fl", "br",
  "tr", "tl", "bl"))
}
\arguments{
\item{show}{\emph{Required, logical} Show color key and density
information? \code{TRUE/FALSE}. Defaults to \code{FALSE}}

\item{title}{\emph{Optional, character} Separate title for legend. 
Defaults to \code{NULL}.}

\item{location}{\emph{Required, character} Location for legend, either 
\code{"fl", "br", "tr", "tl", or "bl"} for "float", "bottom right", 
"top right", "top left", and "bottom left". Defaults to \code{"fl"}, 
which follows the location of the axis labels.}
}
\value{
Modified d3heatmap object
}
\description{
Provides options for modifying the legend for a d3heatmap object
}
