% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmLegend.R
\name{hmLegend}
\alias{hmLegend}
\title{Modify the d3heatmap legend}
\source{
The interface was inspired by \cite{dygraphs}
}
\usage{
hmLegend(d3heatmap, show = TRUE, size, title, location = c("fl", "br", "tr",
  "tl", "bl"), density = c("histogram", "none"), density.fill = NULL)
}
\arguments{
\item{d3heatmap}{\emph{Required} A valid \emph{d3heatmap} object}

\item{show}{\emph{Required, logical} Show color key and density
information? \code{TRUE/FALSE}. Defaults to \code{FALSE}}

\item{size}{\emph{numeric} value indicating the relative size of the key. 
Will multiple the row and column axis sizes. Defaults to 1}

\item{title}{\emph{Optional, character} Separate title for color key (legend). 
Defaults to \code{NULL}.}

\item{location}{\emph{Required, character} Location for legend (color key), either 
\code{"fl", "br", "tr", "tl", or "bl"} for "float", "bottom right", 
"top right", "top left", and "bottom left". Defaults to \code{"fl"}, 
which follows the location of the axis labels.}

\item{density}{\emph{Optional, character} indicating whether to superimpose 
a 'histogram' or no plot ('none') on the color-key (legend).}

\item{density.fill}{\emph{Optional, character} giving the color for the density 
display specified by 'density', defaults to \code{NULL} which will fill the
bars with the same color as the bin}
}
\value{
Modified d3heatmap object
}
\description{
Provides options for modifying the legend for a d3heatmap object
}
\examples{
library(d3heatmap)
library(magrittr)

d3heatmap(mtcars, dendrogram = 'none', scale = 'column', srtCol = 30) \%>\% 
  hmLegend(show = TRUE, title = "Title", location = "tl") 
  
}
\seealso{
\link{heatmap}, \link[gplots]{heatmap.2}
}
